Add support for central net; Defaults still on public relay
