+++ Golem/Golem.cs
@@ -343,8 +343,10 @@ namespace Golem
             }
         }
 
-        public Golem(string golemPath, string? dataDir, ILoggerFactory loggerFactory, Network network)
+        public Golem(string golemPath, string? dataDir, ILoggerFactory loggerFactory, Network network, RelayType net = RelayType.Public)
         {
+            NetConfig.SetEnv(net);
+
             var prov_datadir = dataDir != null ? Path.Combine(dataDir, "provider") : "./provider";
             var yagna_datadir = dataDir != null ? Path.Combine(dataDir, "yagna") : "./yagna";
