+++ GolemLib/net.cs
@@ -1,0 +1,12 @@
+namespace GolemLib;
+
+public enum RelayType
+{
+    // yacn2.dev.golem.network:7477
+    Public,
+    // yacn2a.dev.golem.network:7477
+    Devnet,
+    // 127.0.0.1:16464
+    Local,
+    Central,
+}
